django_app/requirements.txt
django==5.0.0
psycopg2-binary==2.9.8
python-dotenv==1.0.0
langchain==0.3.3
mistralai==1.1.0
pgvector==0.3.5
numpy==1.24.0
langchain-mistralai==0.2.0
langchain-core==0.3.10
langchain-community==0.3.2
faiss-cpu
djangorestframework
django-cors-headers
recipe-scrapers